+++ Src/FluentAssertions/Numeric/SByteAssertions.cs
@@ -1,0 +1,22 @@
+﻿using System.Diagnostics;
+
+namespace FluentAssertions.Numeric
+{
+    /// <summary>
+    /// Contains a number of methods to assert that an <see cref="sbyte"/> is in the expected state.
+    /// </summary>
+    [DebuggerNonUserCode]
+    internal class SByteAssertions : NumericAssertions<sbyte>
+    {
+        internal SByteAssertions(sbyte value)
+            : base(value)
+        {
+        }
+
+        private protected override sbyte? CalculateDifferenceForFailureMessage(sbyte expected)
+        {
+            var difference = (sbyte?)(Subject - expected);
+            return difference != 0 ? difference : null;
+        }
+    }
+}